PDO-944 - Updates for PD to use skbn to move file from and to k8s/cloud storage.

See merge request ping-cloud-private-tenant/ping-cloud-base!297

+++ profiles/aws/pingdirectory/hooks/82-upload-csd-s3.sh
@@ -2,6 +2,8 @@
 
 ${VERBOSE} && set -x
 
+. "${HOOKS_DIR}/utils.lib.sh"
+
 test -f "${STAGING_DIR}/env_vars" && . "${STAGING_DIR}/env_vars"
 
 # Set PATH - since this is executed from within the server process, it may not have all we need on the path
@@ -11,43 +13,24 @@ export PATH="${PATH}:${SERVER_ROOT_DIR}/bin:/usr/local/bin:/usr/sbin:/usr/bin:/s
 test ! -z "${1}" && LOG_ARCHIVE_URL="${1}"
 echo "Uploading to location ${LOG_ARCHIVE_URL}"
 
-# Install AWS CLI if the upload location is S3
-if test "${LOG_ARCHIVE_URL#s3}" == "${LOG_ARCHIVE_URL}"; then
-  echo "Upload location is not S3"
-  exit 0
-elif ! which aws > /dev/null; then
-  echo "Installing AWS CLI"
-  apk --update add python3
-  pip3 install --no-cache-dir --upgrade pip
-  pip3 install --no-cache-dir --upgrade awscli
+if ! cd "${OUT_DIR}"; then
+  echo "Failed to chdir to: ${OUT_DIR}"
+  exit 1
 fi
-
-FORMAT="+%d/%b/%Y:%H:%M:%S %z"
-NOW=$(date "${FORMAT}")
-
-cd "${OUT_DIR}"
 
 collect-support-data --duration 1h
 CSD_OUT=$(find . -name support\*zip -type f | sort | tail -1)
 
-BUCKET_URL_NO_PROTOCOL=${LOG_ARCHIVE_URL#s3://}
-BUCKET_NAME=$(echo ${BUCKET_URL_NO_PROTOCOL} | cut -d/ -f1)
+# Set required environment variables for skbn
+initializeSkbnConfiguration "${LOG_ARCHIVE_URL}"
 
-DIRECTORY_NAME=$(echo ${PING_PRODUCT} | tr '[:upper:]' '[:lower:]')
-echo "Creating directory ${DIRECTORY_NAME} under bucket ${BUCKET_NAME}"
-aws s3api put-object --bucket "${BUCKET_NAME}" --key "${DIRECTORY_NAME}"/
+DST_FILE="${OUT_DIR}/$(basename "${CSD_OUT}")"
 
-if test "${LOG_ARCHIVE_URL}" == */pingdirectory; then
-  TARGET_URL="${LOG_ARCHIVE_URL}"
-else
-  TARGET_URL="${LOG_ARCHIVE_URL}/${DIRECTORY_NAME}"
+echo "Copying: '${DST_FILE}' to '${SKBN_CLOUD_PREFIX}'"
+
+if ! skbnCopy "${SKBN_K8S_PREFIX}/${DST_FILE}" "${SKBN_CLOUD_PREFIX}/${DST_FILE}"; then
+  exit 1
 fi
-
-echo "Uploading "${CSD_OUT}" to ${TARGET_URL} at ${NOW}"
-DST_FILE=$(basename "${CSD_OUT}")
-aws s3 cp "${CSD_OUT}" "${TARGET_URL}/${DST_FILE}"
-
-echo "Upload return code: ${?}"
 
 # Remove the CSD file so it is doesn't fill up the server's filesystem.
 rm -f "${CSD_OUT}"
