+++ profiles/aws/pingdirectory/hooks/90-upload-backup-s3.sh
@@ -3,10 +3,12 @@ set -e
 
 . "${HOOKS_DIR}/utils.lib.sh"
 
+
+# Source generated environment variables 
 test -f "${STAGING_DIR}/env_vars" && . "${STAGING_DIR}/env_vars"
 
-# Install AWS CLI and set required environment variables for AWS S3 bucket
-initializeS3Configuration
+# Set required environment variables for skbn
+initializeSkbnConfiguration
 
 # This is the backup directory on the server
 SERVER_BACKUP_DIR="${OUT_DIR}/backup"
@@ -32,17 +34,23 @@ done
 
 # Zip backup files and append the current timestamp to zip filename
 cd "${SERVER_BACKUP_DIR}"
-DST_FILE="data-`date +%m-%d-%Y.%H.%M.%S`.zip"
-zip -r ${DST_FILE} *
+DST_FILE_TIMESTAMP="data-$(date +%m-%d-%Y.%H.%M.%S).zip"
+zip -r "${DST_FILE_TIMESTAMP}" *
 
-echo "Creating directory ${DIRECTORY_NAME} under bucket ${BUCKET_NAME}"
-aws s3api put-object --bucket "${BUCKET_NAME}" --key "${DIRECTORY_NAME}"/
+UPLOAD_DIR="${mktemp -p}"
 
-echo "Uploading ${SERVER_BACKUP_DIR}/${DST_FILE} to ${TARGET_URL}"
-aws s3 cp "${SERVER_BACKUP_DIR}/${DST_FILE}" "${TARGET_URL}/"
+# Two copy of the backup will be pushed to cloud storage.
+# Make a copy: latest.zip
+DST_FILE_LATEST=latest.zip
+cp "$DST_FILE_TIMESTAMP" "${UPLOAD_DIR}/${DST_FILE_TIMESTAMP}"
+cp "$DST_FILE_TIMESTAMP" "${UPLOAD_DIR}/${DST_FILE_LATEST}"
 
-# Print the filename of the uploaded file to s3
-echo "${DST_FILE}"
+echo "Copying files in '${UPLOAD_DIR}' to '${SKBN_CLOUD_PREFIX}'"
+
+if ! skbnCopy "${SKBN_K8S_PREFIX}/${UPLOAD_DIR}" "${SKBN_CLOUD_PREFIX}/"; then
+  exit 1
+fi
 
 # Cleanup
 rm -rf "${SERVER_BACKUP_DIR}"
+rm -rf "${UPLOAD_DIR}"
